+++ tuxemon/event/actions/get_player_monster.py
@@ -11,7 +11,7 @@ from tuxemon.event.eventaction import EventAction
 from tuxemon.menu.interface import MenuItem
 from tuxemon.monster import Monster
 from tuxemon.session import Session
-from tuxemon.states.monster import MonsterMenuState
+from tuxemon.states.monster_menu import MonsterMenuState
 from tuxemon.tools import compare
 
 logger = logging.getLogger(__name__)
