+++ tuxemon/states/combat/__init__.py
@@ -1,6 +1,3 @@
 # SPDX-License-Identifier: GPL-3.0
 # Copyright (c) 2014-2025 William Edwards <shadowapex@gmail.com>, Benjamin Bean <superman2k5@gmail.com>
-from .combat import CombatState, WaitForInputState
-from .combat_animations import CombatAnimations
 from .combat_menus import CombatTargetMenuState, MainCombatMenuState
-from .combat_menus_park import MainParkMenuState
