+++ tuxemon/states/craft_menu.py
