--- tuxemon/states/items/__init__.py
@@ -1,16 +1,0 @@
-# SPDX-License-Identifier: GPL-3.0
-# Copyright (c) 2014-2025 William Edwards <shadowapex@gmail.com>, Benjamin Bean <superman2k5@gmail.com>
-from __future__ import annotations
-
-from .craft_menu import CraftMenuState
-from .item_menu import ItemMenuState
-from .shop_item import (
-    ShopItemBuyMenuState,
-    ShopItemMenuState,
-    ShopItemSellMenuState,
-)
-from .shop_monster import (
-    ShopMonsterBuyMenuState,
-    ShopMonsterMenuState,
-    ShopMonsterSellMenuState,
-)
