+++ tuxemon/states/monster_item.py
@@ -17,7 +17,7 @@ from tuxemon.menu.menu import PygameMenuState
 from tuxemon.monster import Monster
 from tuxemon.platform.const import buttons
 from tuxemon.platform.events import PlayerInput
-from tuxemon.states.items.item_menu import ItemMenuState
+from tuxemon.states.item_menu import ItemMenuState
 
 
 class MonsterItemState(PygameMenuState):
