+++ tuxemon/states/monster_menu.py
