+++ tuxemon/states/shop_item.py
@@ -9,20 +9,19 @@ from typing import TYPE_CHECKING, ClassVar, Optional
 from pygame.rect import Rect
 
 from tuxemon import prepare, tools
-from tuxemon.item.item import INFINITE_ITEMS
+from tuxemon.item.item import INFINITE_ITEMS, Item
+from tuxemon.item.shop_utils import (
+    TransactionManager,
+    calc_internal_rect,
+    filter_inventory,
+    generate_label,
+)
 from tuxemon.menu.interface import MenuItem
 from tuxemon.menu.menu import Menu
 from tuxemon.menu.quantity import QuantityAndCostMenu, QuantityAndPriceMenu
-from tuxemon.monster import Monster
 from tuxemon.platform.const import buttons
 from tuxemon.platform.events import PlayerInput
 from tuxemon.sprite import Sprite
-from tuxemon.states.items.shop_utils import (
-    TransactionManager,
-    calc_internal_rect,
-    filter_party,
-    generate_label,
-)
 from tuxemon.ui.paginator import Paginator
 from tuxemon.ui.text import TextArea
 
@@ -31,9 +30,9 @@ if TYPE_CHECKING:
     from tuxemon.npc import NPC
 
 
-class ShopMonsterMenuState(Menu[Monster]):
+class ShopItemMenuState(Menu[Item]):
 
-    name: ClassVar[str] = "ShopMonsterMenuState"
+    name: ClassVar[str] = "ShopItemMenuState"
     draw_borders = False
 
     def __init__(
@@ -45,17 +44,17 @@ class ShopMonsterMenuState(Menu[Monster]):
     ) -> None:
         super().__init__()
 
-        # this sprite is used to display the monster
+        # this sprite is used to display the item
         self.item_center = self.rect.width * 0.164, self.rect.height * 0.13
-        self.monster_sprite = Sprite()
-        self.sprites.add(self.monster_sprite)
+        self.item_sprite = Sprite()
+        self.sprites.add(self.item_sprite)
 
         self.menu_items.line_spacing = tools.scale(7)
         self.current_page = 0
         self.total_pages = 0
-        self.inventory: list[Monster] = []
+        self.inventory: list[Item] = []
 
-        # this is the area where the monster description is displayed
+        # this is the area where the item description is displayed
         rect = self.client.screen.get_rect()
         rect.top = tools.scale(106)
         rect.left = tools.scale(3)
@@ -81,80 +80,71 @@ class ShopMonsterMenuState(Menu[Monster]):
     def calc_internal_rect(self) -> Rect:
         return calc_internal_rect(self.rect)
 
-    def is_valid_entry(self, monster: Optional[Monster]) -> bool:
-        """Check if the selected monster is valid for purchase or sale."""
-        if not monster:
+    def is_valid_entry(self, item: Optional[Item]) -> bool:
+        """Check if the selected item is valid for purchase or sale."""
+        if not item:
             return False
         if self.buyer.is_player:
-            _, _, price = generate_label(monster, self.economy, 1)
+            _, _, price = generate_label(item, self.economy, 1)
             wallet = self.buyer_manager.get_money()
-            key = f"{self.economy.model.slug}:{monster.slug}"
+            key = f"{self.economy.model.slug}:{item.slug}"
             qty = self.buyer.game_variables.get(key, 0)
             if price > wallet or qty == 0:
-                return False
-        if self.seller.is_player:
-            if self.seller.party.party_size == 1:
                 return False
         return True
 
     def on_menu_selection_change(self) -> None:
         """Called when menu selection changes."""
-        monster = self.get_selected_item()
-        if monster:
-            image = monster.game_object.get_sprite("front")
+        item = self.get_selected_item()
+        if item:
+            image = item.game_object.surface
             assert image
-            self.monster_sprite.image = image.image
-            self.monster_sprite.rect = image.image.get_rect(
-                center=self.image_center
-            )
-            if monster.description:
-                self.dialog.alert(monster.description, dialog_speed="max")
+            self.item_sprite.image = image
+            self.item_sprite.rect = image.get_rect(center=self.image_center)
+            if item.description:
+                self.dialog.alert(item.description, dialog_speed="max")
 
-    def generate_monster_label(
+    def generate_label(
         self,
-        monster: Monster,
+        item: Item,
         qty: Optional[int] = None,
         seller_mode: bool = False,
     ) -> tuple[str, str, int]:
-        """Generate the label for shop monsters, handling both buyer and seller modes."""
-        return generate_label(monster, self.economy, qty, seller_mode)
+        """Generate the label for shop items, handling both buyer and seller modes."""
+        return generate_label(item, self.economy, qty, seller_mode)
 
-    def _populate_menu_monsters(
-        self, inventory: list[Monster]
-    ) -> Generator[MenuItem[Monster], None, None]:
-        for monster in inventory:
+    def _populate_menu_items(
+        self, inventory: list[Item]
+    ) -> Generator[MenuItem[Item], None, None]:
+        for item in inventory:
             if self.buyer.is_player:
-                key = f"{self.economy.model.slug}:{monster.slug}"
+                key = f"{self.economy.model.slug}:{item.slug}"
                 qty = self.buyer.game_variables.get(key, 0)
-                label, discount, price = self.generate_monster_label(
-                    monster, qty
-                )
+                label, discount, price = self.generate_label(item, qty)
                 fg = (
                     self.unavailable_color_shop
                     if price > self.buyer_manager.get_money()
                     else None
                 )
                 image = self.shadow_text(label, fg=fg)
-                menu_monster = MenuItem(
-                    image, monster.name, monster.description, monster
-                )
-                yield menu_monster
-                menu_monster.metadata["price"] = price
-                self.add(menu_monster)
+                menu_item = MenuItem(image, item.name, item.description, item)
+                yield menu_item
+                menu_item.metadata["price"] = price
+                self.add(menu_item)
             elif self.seller.is_player:
-                label, discount, cost = self.generate_monster_label(
-                    monster, qty=None, seller_mode=True
+                label, discount, cost = self.generate_label(
+                    item, qty=None, seller_mode=True
                 )
                 image = self.shadow_text(label)
-                menu_monster = MenuItem(
-                    image, monster.name, monster.description, monster
-                )
-                yield menu_monster
-                menu_monster.metadata["cost"] = cost
-                self.add(menu_monster)
+                menu_item = MenuItem(image, item.name, item.description, item)
+                yield menu_item
+                menu_item.metadata["cost"] = cost
+                self.add(menu_item)
 
-    def initialize_items(self) -> Generator[MenuItem[Monster], None, None]:
-        self.inventory = filter_party(self.buyer, self.seller, self.economy)
+    def initialize_items(self) -> Generator[MenuItem[Item], None, None]:
+        self.inventory = filter_inventory(
+            self.buyer, self.seller, self.economy
+        )
         if not self.inventory:
             return
 
@@ -165,15 +155,17 @@ class ShopMonsterMenuState(Menu[Monster]):
         )
 
         paged_inventory = self.paginator.paginate(self.current_page)
-        yield from self._populate_menu_monsters(paged_inventory)
+        yield from self._populate_menu_items(paged_inventory)
 
     def reload_shop(self) -> None:
         self.clear()
-        self.inventory = filter_party(self.buyer, self.seller, self.economy)
+        self.inventory = filter_inventory(
+            self.buyer, self.seller, self.economy
+        )
 
         paged_inventory = self.paginator.paginate(self.current_page)
         # Force generator execution
-        list(self._populate_menu_monsters(paged_inventory))
+        list(self._populate_menu_items(paged_inventory))
 
         self.selected_index = (
             min(self.selected_index, len(self.menu_items) - 1)
@@ -201,24 +193,24 @@ class ShopMonsterMenuState(Menu[Monster]):
         return None
 
 
-class ShopMonsterBuyMenuState(ShopMonsterMenuState):
-    """State for buying monsters."""
+class ShopItemBuyMenuState(ShopItemMenuState):
+    """State for buying items."""
 
-    name: ClassVar[str] = "ShopMonsterBuyMenuState"
+    name: ClassVar[str] = "ShopItemBuyMenuState"
 
-    def on_menu_selection(self, menu_monster: MenuItem[Monster]) -> None:
-        monster = menu_monster.game_object
-        price: int = menu_monster.metadata.get("price", 1)
-        label = f"{self.economy.model.slug}:{monster.slug}"
+    def on_menu_selection(self, menu_item: MenuItem[Item]) -> None:
+        item = menu_item.game_object
+        price: int = menu_item.metadata.get("price", 1)
+        label = f"{self.economy.model.slug}:{item.slug}"
 
-        def buy_monster(quantity: int) -> None:
-            self.transaction_manager.buy_monster(
-                self.buyer, monster, quantity, label, price
+        def buy_item(quantity: int) -> None:
+            self.transaction_manager.buy_item(
+                self.buyer, item, quantity, label, price
             )
             self.reload_items()
             if (
                 self.seller.shop_inventory
-                and not self.seller.shop_inventory.has_monster(monster.slug)
+                and not self.seller.shop_inventory.has_item(item.slug)
             ):
                 self.on_menu_selection_change()
 
@@ -233,7 +225,7 @@ class ShopMonsterBuyMenuState(ShopMonsterMenuState):
 
         self.client.push_state(
             QuantityAndPriceMenu(
-                callback=partial(buy_monster),
+                callback=partial(buy_item),
                 max_quantity=max_quantity,
                 quantity=1,
                 shrink_to_items=True,
@@ -242,33 +234,35 @@ class ShopMonsterBuyMenuState(ShopMonsterMenuState):
         )
 
 
-class ShopMonsterSellMenuState(ShopMonsterMenuState):
-    """State for selling monsters."""
+class ShopItemSellMenuState(ShopItemMenuState):
+    """State for selling items."""
 
-    name: ClassVar[str] = "ShopMonsterSellMenuState"
+    name: ClassVar[str] = "ShopItemSellMenuState"
 
-    def on_menu_selection(self, menu_monster: MenuItem[Monster]) -> None:
-        monster = menu_monster.game_object
-        metadata_cost = menu_monster.metadata.get("cost")
-        basic_cost = self.economy.lookup_item_field(monster.slug, "cost")
+    def on_menu_selection(self, menu_item: MenuItem[Item]) -> None:
+        item = menu_item.game_object
+        metadata_cost = menu_item.metadata.get("cost")
+        basic_cost = self.economy.lookup_item_field(item.slug, "cost")
 
         if metadata_cost is not None:
             cost = metadata_cost
         elif basic_cost:
             cost = basic_cost
         else:
-            cost = round(monster.hp * self.economy.model.resale_multiplier)
+            cost = round(item.cost * self.economy.model.resale_multiplier)
 
-        def sell_monster(quantity: int) -> None:
-            self.transaction_manager.sell_monster(self.seller, monster, cost)
+        def sell_item(quantity: int) -> None:
+            self.transaction_manager.sell_item(
+                self.seller, item, quantity, cost
+            )
             self.reload_items()
-            if not self.seller.party.has_monster(monster):
+            if not self.seller.items.has_item(item.slug):
                 self.on_menu_selection_change()
 
         self.client.push_state(
             QuantityAndCostMenu(
-                callback=partial(sell_monster),
-                max_quantity=1,
+                callback=partial(sell_item),
+                max_quantity=item.quantity,
                 quantity=1,
                 shrink_to_items=True,
                 cost=cost,
