+++ tuxemon/states/world/__init__.py
@@ -1,4 +1,3 @@
 # SPDX-License-Identifier: GPL-3.0
 # Copyright (c) 2014-2025 William Edwards <shadowapex@gmail.com>, Benjamin Bean <superman2k5@gmail.com>
-from .world_menus import WorldMenuState
 from .worldstate import WorldState
